PLAT-394: split workflow views
